fix: show path import snippets on parent nodes
